feat: Student Member에서 분리
