+++ app/components/screens/member/Member.tsx
@@ -15,6 +15,7 @@ import {
   heightPercentage,
   widthPercentage,
 } from '../../../utils/constant/common/design/Responsive';
+import Student from '../../common/Student';
 
 const styles = StyleSheet.create({
   root: {
@@ -91,43 +92,11 @@ const styles = StyleSheet.create({
     marginTop: heightPercentage(15),
     marginLeft: widthPercentage(20),
   },
-  item: {
-    width: widthPercentage(307),
-    height: heightPercentage(50),
-    marginTop: 5,
-    marginHorizontal: 14,
-  },
-  name: {
-    width: widthPercentage(44),
-    height: heightPercentage(24),
-    fontFamily: 'NotoSansKR-Regular',
-    fontSize: fontPercentage(16),
-    textAlign: 'left',
-    color: '#000000',
-  },
   itemSeparator: {
     marginLeft: widthPercentage(14),
     height: heightPercentage(1),
     width: widthPercentage(307),
     backgroundColor: '#c2c2c2',
-  },
-  major: {
-    width: widthPercentage(55),
-    height: heightPercentage(17),
-    fontFamily: 'NotoSansKR-Regular',
-    fontSize: fontPercentage(12),
-    color: '#808080',
-  },
-  studentCode: {
-    width: widthPercentage(67),
-    height: heightPercentage(17),
-    fontFamily: 'NotoSansKR-Regular',
-    fontSize: fontPercentage(12),
-    marginLeft: widthPercentage(185),
-    color: '#808080',
-  },
-  itemSection: {
-    flexDirection: 'row',
   },
 });
 
@@ -183,23 +152,6 @@ const DATA = [
     studentCode: '2021086326',
   },
 ];
-const Student = ({
-  name,
-  major,
-  studentCode,
-}: {
-  name: string;
-  major: string;
-  studentCode: number;
-}) => (
-  <View style={styles.item}>
-    <Text style={styles.name}>{name}</Text>
-    <View style={styles.itemSection}>
-      <Text style={styles.major}>{major}</Text>
-      <Text style={styles.studentCode}>{studentCode}</Text>
-    </View>
-  </View>
-);
 const searchIcon = require('../../../assets/images/searchIcon.png');
 
 const renderItem: ({item}: {item: Array}) => any = ({item}) => (
